+++ CHANGELOG.md
@@ -2,13 +2,26 @@
 
 All notable changes to Fog of Walk are documented here.
 
+## [0.4.0] - 2026-08-31
+
+### Changed
+
+- add landing page (#32)
+
+## [0.3.0] - 2026-08-31
+
+### Changed
+
+- refactor map view, fix several issues related to map & fog processing (#31)
+- minor saved points improvements (#30)
+
 ## [0.2.1] - 2026-08-30
 
 ### Changed
 
 - add uuid fallback
 - improve release generation logic
-- fix mobile issues in saved point edit dialog
+- fix mobile issues in saved point edit dialog (part of them)
 
 ## [0.2.0] - 2026-08-30